almost done - 4

+++ part4/controllers/blogs.js
@@ -1,13 +1,26 @@
 const blogsRouter = require('express').Router()
-//const jwt = require('jsonwebtoken')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 
 
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+      return authorization.substring(7)
+    }
+    return null
+  }
 
-
-blogsRouter.post('/', (request, response) => {
-    var blog
+blogsRouter.post('/', async (request, response) => {
+    var blog =undefined
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    //const user = await User.findById(decodedToken.id)
     if (!request.body.title && !request.body.url)
         return response.status(400).json({ error: 'title and url is missing' })
 
@@ -21,32 +34,45 @@ blogsRouter.post('/', (request, response) => {
     }else{
         blog = new Blog(request.body)
     }
-    
-    
-    blog
-      .save()
-      .then(result => {
-        response.status(201).json(result)
-      })
+    const user = await User.findById(decodedToken.id)
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    user.save()
+    response.set(200).json(savedBlog)
   })
   
-blogsRouter.get('/:id', (request, response) => {
-    Blog.findById(request.params.id)
-      .then(blog => {
+blogsRouter.get('/:id',async  (request, response) => {
+    const blog =await Blog.findById(request.params.id)
         if (blog) {
-          response.json(blog)
+          response.json(blog.toJSON())
         } else {
           response.status(404).end()
         }
-      })
+
   })
 
 blogsRouter.delete('/:id', async (request, response) => {
-      await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
+
+
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    const blog = await Blog.findById(request.params.id)
+
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } else {
+        return response.status(401).json({error: "you do not have permission to delete this blog"})
+      }
 })
 
-blogsRouter.put('/:id', (request, response) => {
+blogsRouter.put('/:id',async  (request, response) => {
   
     const blog = {
       title: request.body.title,
@@ -54,14 +80,16 @@ blogsRouter.put('/:id', (request, response) => {
       url: request.body.url,
       likes: request.body.likes,
     }
-  
-    Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-      .then(updatedBlog => {
-        response.json(updatedBlog)
-      })
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+    if (updatedBlog) {
+        //TODO
+        response.status(200).json(updatedBlog.toJSON())
+      } else {
+        response.status(404).end()
+      }
   })
 
-  
+
 blogsRouter.get('/', (request, response) => {
     Blog
       .find({})
